Update role so we're not reloading the daemon in excess
The systemd-networkd role was reloading the systemd daemon on every run
this change removes that as this role will not need to modify anything
that requires a daemon reload.

Network facts will be gathered after this role executes which will
ensure any state changes are picked up as soon as they're available.

Change-Id: Ib3efa6a1662da48d0f58e00ed77f725e1a950edb
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>