Add repos only when there're packages for install
At the moment we allow not to install packages by providing
systemd_networkd_distro_packages as an empty list. When that happens
it makes sense also to skip all tasks related to adding repositories
when no packages are going to be installed.

Not installing packages might be useful for containers creation or
in situations when there is no networking yet, but we know that
systemd-networkd is already present.

Change-Id: I5e12cb0eb07b5eb8dcca462dbe8239a926ef80e6