Liberalize static_route defenition
At the moment a very limited amount of options are supported for
static_routes defenition, which might be a limiting factor in some more
complex solutions.

It is proposed to let operators define `static_routes` with parameters
that are available in systemd-networkd [1] directly, which should
provide required flexbility and not require us additional maintenance
and regular updates of the code.

[1] https://www.freedesktop.org/software/systemd/man/latest/systemd.network.html#%5BRoute%5D%20Section%20Options

Change-Id: I71f81cb497ae8eeb0fab2bc6f9f5eae120d10775