Do not try to configure resolved when it's not available
While having `systemd_resolved_available` variable which controls if
systemd-resolved will be restarted in handlers.

However, way better approach is to prevent resolved from being
configured at the first place, which will result in skipping handler
execution on it's own.

With that we add control over systemd-resolved package installed with
the same `systemd_resolved_available` variable, so that package
installation is omited when variable is set to False.

Change-Id: Ib2471850386bb125a92c02fb21874e100a03459d