Install role pre-requisite packages
Historically we were assuming some subsystems like udev or dbus
being part of the core images and never handled their installation.
However such subsystems are a hard requirement for role to succeed
and networkd to operate correctly.

To ensure all required bits in place, we've adding installation of dbus
and udev to the role in order to prevent failures on very minimal images

Change-Id: Ie77a7064d17e6aef0be16b8a5d9be5cc121b665b