Only restart non-networkd services when the role is configured to install them
The variable systemd_networkd_distro_packages can be overridden
externally to be an empty list, to ensure the minimum number of
tasks executes in this role. This might happen when the target
is know to already have the pre-requisite packages present because
they have been built into a container or VM image previously.

This patch adjusts the non-networkd service start handler to only
run when the distro package list is not empty.

Change-Id: I6bfb5b66dfef090e3de0d90f726a06e290c09bc0